+++ internal/resource/manage/kubectl/diff_test.go
@@ -99,7 +99,7 @@ func TestDiffManagerApply(t *testing.T) {
 				mk.On("EncodeObjects", mock.Anything, expK8sResources).Once().Return([]byte("test"), nil)
 
 				exp := expCmdMatcher(
-					[]string{"kubectl", "diff", "--force-conflicts=true", "--server-side=true", "-f", "-"},
+					[]string{"kubectl", "diff", "--force-conflicts=true", "--server-side=true", "--filename", "-"},
 					"test",
 				)
 				mc.On("Run", mock.MatchedBy(exp)).Once().Return(nil)
@@ -115,7 +115,7 @@ func TestDiffManagerApply(t *testing.T) {
 				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test"), nil)
 
 				exp := expCmdMatcher(
-					[]string{"whatever", "diff", "--force-conflicts=true", "--server-side=true", "-f", "-"},
+					[]string{"whatever", "diff", "--force-conflicts=true", "--server-side=true", "--filename", "-"},
 					"test",
 				)
 				mc.On("Run", mock.MatchedBy(exp)).Once().Return(nil)
@@ -131,7 +131,7 @@ func TestDiffManagerApply(t *testing.T) {
 				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test"), nil)
 
 				exp := expCmdMatcher(
-					[]string{"kubectl", "diff", "--context", "whatever", "--force-conflicts=true", "--server-side=true", "-f", "-"},
+					[]string{"kubectl", "diff", "--context", "whatever", "--force-conflicts=true", "--server-side=true", "--filename", "-"},
 					"test",
 				)
 				mc.On("Run", mock.MatchedBy(exp)).Once().Return(nil)
@@ -147,7 +147,7 @@ func TestDiffManagerApply(t *testing.T) {
 				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test"), nil)
 
 				exp := expCmdMatcher(
-					[]string{"kubectl", "diff", "--kubeconfig", "whatever", "--force-conflicts=true", "--server-side=true", "-f", "-"},
+					[]string{"kubectl", "diff", "--kubeconfig", "whatever", "--force-conflicts=true", "--server-side=true", "--filename", "-"},
 					"test",
 				)
 				mc.On("Run", mock.MatchedBy(exp)).Once().Return(nil)
@@ -163,7 +163,7 @@ func TestDiffManagerApply(t *testing.T) {
 				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test"), nil)
 
 				exp := expCmdMatcher(
-					[]string{"kubectl", "diff", "--force-conflicts=false", "--server-side=true", "-f", "-"},
+					[]string{"kubectl", "diff", "--force-conflicts=false", "--server-side=true", "--filename", "-"},
 					"test",
 				)
 				mc.On("Run", mock.MatchedBy(exp)).Once().Return(nil)
@@ -179,7 +179,7 @@ func TestDiffManagerApply(t *testing.T) {
 				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test"), nil)
 
 				exp := expCmdMatcher(
-					[]string{"kubectl", "diff", "--force-conflicts=true", "--field-manager", "whatever", "--server-side=true", "-f", "-"},
+					[]string{"kubectl", "diff", "--force-conflicts=true", "--field-manager", "whatever", "--server-side=true", "--filename", "-"},
 					"test",
 				)
 				mc.On("Run", mock.MatchedBy(exp)).Once().Return(nil)
@@ -200,6 +200,7 @@ func TestDiffManagerApply(t *testing.T) {
 			// Prepare.
 			test.config.Out = ioutil.Discard
 			test.config.YAMLEncoder = menc
+			test.config.YAMLDecoder = &kubectlmock.K8sObjectDecoder{}
 			test.config.CmdRunner = mcmd
 			manager, err := kubectl.NewDiffManager(test.config)
 			require.NoError(err)
@@ -222,19 +223,62 @@ func TestDiffManagerDelete(t *testing.T) {
 	tests := map[string]struct {
 		config    kubectl.DiffManagerConfig
 		resources []model.Resource
-		mock      func(mk *kubectlmock.K8sObjectEncoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager)
+		mock      func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager)
 		expErr    bool
 	}{
 		"Not having resources, shouldn't execute anything.": {
 			resources: []model.Resource{},
-			mock:      func(mk *kubectlmock.K8sObjectEncoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {},
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+			},
+		},
+
+		"Having an error while encoding objects to get the latest state from the resources in the server, should fail.": {
+			resources: []model.Resource{
+				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
+			},
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return(nil, fmt.Errorf("whatever"))
+			},
+			expErr: true,
+		},
+
+		"Having an error while running kubctl get to get the latest state from the resources in the server, should fail.": {
+			resources: []model.Resource{
+				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
+			},
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
+				mc.On("Run", mock.Anything).Once().Return(fmt.Errorf("whatever"))
+			},
+			expErr: true,
+		},
+
+		"Having an error while decodding latest received state from the server, should fail.": {
+			resources: []model.Resource{
+				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
+			},
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
+				mc.On("Run", mock.Anything).Once().Return(nil)
+				mkd.On("DecodeObjects", mock.Anything, mock.Anything).Once().Return(nil, fmt.Errorf("whatever"))
+			},
+			expErr: true,
 		},
 
 		"Having an error while creating the tmp dir should stop execution and fail.": {
 			resources: []model.Resource{
 				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
 			},
-			mock: func(mk *kubectlmock.K8sObjectEncoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
+				mc.On("Run", mock.Anything).Once().Return(nil)
+				mkd.On("DecodeObjects", mock.Anything, mock.Anything).Once().Return([]model.K8sObject{newK8sObject("test1", "ns1")}, nil)
+
+				// Diff part.
 				mfs.On("TempDir", mock.Anything, mock.Anything).Once().Return("", fmt.Errorf("whatever"))
 			},
 			expErr: true,
@@ -244,11 +288,17 @@ func TestDiffManagerDelete(t *testing.T) {
 			resources: []model.Resource{
 				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
 			},
-			mock: func(mk *kubectlmock.K8sObjectEncoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
+				mc.On("Run", mock.Anything).Once().Return(nil)
+				mkd.On("DecodeObjects", mock.Anything, mock.Anything).Once().Return([]model.K8sObject{newK8sObject("test1", "ns1")}, nil)
+
+				// Diff part.
 				mfs.On("TempDir", mock.Anything, mock.Anything).Once().Return("", nil)
 				mfs.On("RemoveAll", mock.Anything).Once().Return(nil)
 
-				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return(nil, fmt.Errorf("whatever"))
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return(nil, fmt.Errorf("whatever"))
 			},
 			expErr: true,
 		},
@@ -257,27 +307,51 @@ func TestDiffManagerDelete(t *testing.T) {
 			resources: []model.Resource{
 				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
 			},
-			mock: func(mk *kubectlmock.K8sObjectEncoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
+				mc.On("Run", mock.Anything).Once().Return(nil)
+				mkd.On("DecodeObjects", mock.Anything, mock.Anything).Once().Return([]model.K8sObject{newK8sObject("test1", "ns1")}, nil)
+
+				// Diff part.
 				mfs.On("TempDir", mock.Anything, mock.Anything).Once().Return("", nil)
 				mfs.On("RemoveAll", mock.Anything).Once().Return(nil)
 
-				mk.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
+				mke.On("EncodeObjects", mock.Anything, mock.Anything).Once().Return([]byte("test1"), nil)
 				mfs.On("WriteFile", mock.Anything, mock.Anything, mock.Anything).Once().Return(fmt.Errorf("whatever"))
 			},
 			expErr: true,
 		},
 
-		"Having resources should apply correctly.": {
+		"Having resources should delete correctly.": {
 			resources: []model.Resource{
 				{Name: "test1", K8sObject: newK8sObject("test1", "ns1")},
 				{Name: "test2", K8sObject: newK8sObject("test2", "ns2")},
+				{Name: "test3", K8sObject: newK8sObject("test3", "ns3")},
 			},
-			mock: func(mk *kubectlmock.K8sObjectEncoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+			mock: func(mke *kubectlmock.K8sObjectEncoder, mkd *kubectlmock.K8sObjectDecoder, mc *kubectlmock.CmdRunner, mfs *kubectlmock.FSManager) {
+				expK8sResources1 := newK8sObject("test1", "ns1")
+				expK8sResources2 := newK8sObject("test2", "ns2")
+				expK8sResources3 := newK8sObject("test3", "ns3")
+				expK8sResources := []model.K8sObject{expK8sResources1, expK8sResources2, expK8sResources3}
+
+				// Getting server state part.
+				mke.On("EncodeObjects", mock.Anything, expK8sResources).Once().Return([]byte("test1"), nil)
+
+				expCmdKubectl := expCmdMatcher(
+					[]string{"kubectl", "get", "--ignore-not-found=true", "--output", "yaml", "--filename", "-"},
+					"test1",
+				)
+				mc.On("Run", mock.MatchedBy(expCmdKubectl)).Once().Return(nil)
+
+				// Emulate server has returned one less
+				mkd.On("DecodeObjects", mock.Anything, mock.Anything).Once().Return([]model.K8sObject{expK8sResources1, expK8sResources2}, nil)
+
+				// Diff part.
 				mfs.On("TempDir", "", "KAHOY-").Once().Return("/tmp/KAHOY-42", nil)
 				mfs.On("RemoveAll", "/tmp/KAHOY-42").Once().Return(nil)
 
-				expK8sResources1 := []model.K8sObject{newK8sObject("test1", "ns1")}
-				mk.On("EncodeObjects", mock.Anything, expK8sResources1).Once().Return([]byte("test1"), nil)
+				mke.On("EncodeObjects", mock.Anything, []model.K8sObject{expK8sResources1}).Once().Return([]byte("test1"), nil)
 				mfs.On("WriteFile", "/tmp/KAHOY-42/apps.v1.Deployment.ns1.test1", []byte("test1"), mock.Anything).Once().Return(nil)
 				expCmd1 := expCmdMatcher(
 					[]string{"diff", "-u", "-N", "/tmp/KAHOY-42/apps.v1.Deployment.ns1.test1", "-"},
@@ -285,8 +359,7 @@ func TestDiffManagerDelete(t *testing.T) {
 				)
 				mc.On("Run", mock.MatchedBy(expCmd1)).Once().Return(nil)
 
-				expK8sResources2 := []model.K8sObject{newK8sObject("test2", "ns2")}
-				mk.On("EncodeObjects", mock.Anything, expK8sResources2).Once().Return([]byte("test2"), nil)
+				mke.On("EncodeObjects", mock.Anything, []model.K8sObject{expK8sResources2}).Once().Return([]byte("test2"), nil)
 				mfs.On("WriteFile", "/tmp/KAHOY-42/apps.v1.Deployment.ns2.test2", []byte("test2"), mock.Anything).Once().Return(nil)
 				expCmd2 := expCmdMatcher(
 					[]string{"diff", "-u", "-N", "/tmp/KAHOY-42/apps.v1.Deployment.ns2.test2", "-"},
@@ -305,13 +378,15 @@ func TestDiffManagerDelete(t *testing.T) {
 
 			// Mocks.
 			menc := &kubectlmock.K8sObjectEncoder{}
+			mdec := &kubectlmock.K8sObjectDecoder{}
 			mcmd := &kubectlmock.CmdRunner{}
 			mfs := &kubectlmock.FSManager{}
-			test.mock(menc, mcmd, mfs)
+			test.mock(menc, mdec, mcmd, mfs)
 
 			// Prepare.
 			test.config.Out = ioutil.Discard
 			test.config.YAMLEncoder = menc
+			test.config.YAMLDecoder = mdec
 			test.config.CmdRunner = mcmd
 			test.config.FSManager = mfs
 			manager, err := kubectl.NewDiffManager(test.config)